src: remove coded_by tag
We had a tag by codegea (https://codegea.com/), pointing out that the
coded the Astro migration. Initially, it was fine by us; however, a
volunteer pointed out that I would make a conflict of interest to let
this tag sit here.

Oh, man. All that is to want to have some sponsors.